move BellNoneIcon to toolbar
