+++ frontend/src/modules/app/toolbar/icons/index.ts
@@ -1,1 +1,2 @@
-export * from './BellIcon';
+export { default as BellIcon } from './BellIcon';
+export { default as BellNoneIcon } from './BellNoneIcon';
